Major rework of the grub2 PKGBUILD and upgrade to version 2.00.
Also merged back the grub2-efi-x64 package into this PKGBUILD.
The apparent reason for Arch Linux to split it in the first seems to be the ability to break the build on i686 machines for grub-efi-x64 only.
I think this is wrong: the right way to do this is make PKGBUILD ready to build and package it, but actually do it only if the machine architecture is 'x86_64'.