Changing grub2.install again, so that it restores the old configuration also when the package is upgraded.
That is because grub2 1.99 did own /boot/grub/grub.cfg, so upgrading to grub2 2.00 would result in that file being moved to its .pacsave entry.
To avoid this, which would break systems, I'll restore the old configuration, so that the configuration on disk and the bootloader in the MBR should match.
Any other choice (automatically run grub-mkconfig, or automatically run grub-install) is conceptually wrong and could lead to unwanted results or broken systems.