openblas: do not provide blas
Reasoning: We explicitly link against openblas and do not provide
a symlink for libblas. This makes substituting the blas library
impossible, but this is intentional: As a small distribution,
there is little point in supporting multiple blas versions.
Furthermore, anyone who wants maximal performance should not
only use an optimized blas library, but also compile its consumer
against it.